+++ Docs/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added monitoring of chats in OpenTgResearcherConsole
 - Added chat monitoring settings to OpenTgResearcherConsole
 - Added logging of the application OpenTgResearcherConsole to file
+- Added keyword search in chats in OpenTgResearcherConsole
 
 ## [0.6.300] - 2025-06-09
 ### Changed